+++ packages/vite/src/node/plugins/index.ts
@@ -1,6 +1,10 @@
 import aliasPlugin, { type ResolverFunction } from '@rollup/plugin-alias'
 import type { ObjectHook } from 'rolldown'
-import { viteAliasPlugin as nativeAliasPlugin } from 'rolldown/experimental'
+import {
+  viteAliasPlugin as nativeAliasPlugin,
+  viteJsonPlugin as nativeJsonPlugin,
+  viteWasmFallbackPlugin as nativeWasmFallbackPlugin,
+} from 'rolldown/experimental'
 import type { PluginHookUtils, ResolvedConfig } from '../config'
 import {
   type HookHandler,
@@ -8,15 +12,14 @@ import {
   type PluginWithRequiredHook,
 } from '../plugin'
 import { watchPackageDataPlugin } from '../packages'
-import { jsonPlugin } from './json'
-import { oxcResolvePlugin, resolvePlugin } from './resolve'
+import { oxcResolvePlugin } from './resolve'
 import { optimizedDepsPlugin } from './optimizedDeps'
 import { importAnalysisPlugin } from './importAnalysis'
 import { cssAnalysisPlugin, cssPlugin, cssPostPlugin } from './css'
 import { assetPlugin } from './asset'
 import { clientInjectionsPlugin } from './clientInjections'
 import { buildHtmlPlugin, htmlInlineProxyPlugin } from './html'
-import { wasmFallbackPlugin, wasmHelperPlugin } from './wasm'
+import { wasmHelperPlugin } from './wasm'
 import { modulePreloadPolyfillPlugin } from './modulePreloadPolyfill'
 import { webWorkerPlugin } from './worker'
 import { preAliasPlugin } from './preAlias'
@@ -47,7 +50,6 @@ export async function resolvePlugins(
     ? await (await import('../build')).resolveBuildPlugins(config)
     : { pre: [], post: [] }
   const { modulePreload } = config.build
-  const enableNativePlugin = config.nativePluginEnabledLevel >= 0
   const enableNativePluginV1 = config.nativePluginEnabledLevel >= 1
 
   return [
@@ -76,45 +78,33 @@ export async function resolvePlugins(
     modulePreload !== false && modulePreload.polyfill
       ? modulePreloadPolyfillPlugin(config)
       : null,
-    ...(enableNativePlugin
-      ? oxcResolvePlugin(
-          {
-            root: config.root,
-            isProduction: config.isProduction,
-            isBuild,
-            packageCache: config.packageCache,
-            asSrc: true,
-            optimizeDeps: true,
-            externalize: true,
-            legacyInconsistentCjsInterop: config.legacy?.inconsistentCjsInterop,
-          },
-          isWorker
-            ? { ...config, consumer: 'client', optimizeDepsPluginNames: [] }
-            : undefined,
-        )
-      : [
-          resolvePlugin({
-            root: config.root,
-            isProduction: config.isProduction,
-            isBuild,
-            packageCache: config.packageCache,
-            asSrc: true,
-            optimizeDeps: true,
-            externalize: true,
-          }),
-        ]),
+    ...oxcResolvePlugin(
+      {
+        root: config.root,
+        isProduction: config.isProduction,
+        isBuild,
+        packageCache: config.packageCache,
+        asSrc: true,
+        optimizeDeps: true,
+        externalize: true,
+        legacyInconsistentCjsInterop: config.legacy?.inconsistentCjsInterop,
+      },
+      isWorker
+        ? { ...config, consumer: 'client', optimizeDepsPluginNames: [] }
+        : undefined,
+    ),
     htmlInlineProxyPlugin(config),
     cssPlugin(config),
     esbuildBannerFooterCompatPlugin(config),
     config.oxc !== false ? oxcPlugin(config) : null,
-    jsonPlugin(config.json, isBuild, enableNativePluginV1),
-    wasmHelperPlugin(config),
+    nativeJsonPlugin({ ...config.json, minify: isBuild }),
+    wasmHelperPlugin(),
     webWorkerPlugin(config),
     assetPlugin(config),
 
     ...normalPlugins,
 
-    wasmFallbackPlugin(config),
+    nativeWasmFallbackPlugin(),
     definePlugin(config),
     cssPostPlugin(config),
     isBundled && buildHtmlPlugin(config),