+++ packages/vite/src/node/plugins/wasm.ts
@@ -1,13 +1,8 @@
 import MagicString from 'magic-string'
-import { exactRegex } from '@rolldown/pluginutils'
-import {
-  viteWasmFallbackPlugin as nativeWasmFallbackPlugin,
-  viteWasmHelperPlugin as nativeWasmHelperPlugin,
-} from 'rolldown/experimental'
+import { exactRegex } from 'rolldown/filter'
 import { createToImportMetaURLBasedRelativeRuntime } from '../build'
 import type { Plugin } from '../plugin'
 import { cleanUrl } from '../../shared/utils'
-import type { ResolvedConfig } from '..'
 import { assetUrlRE, fileToUrl } from './asset'
 
 const wasmHelperId = '\0vite/wasm-helper.js'
@@ -76,13 +71,7 @@ const instantiateFromFile = async (
 
 const instantiateFromFileCode = instantiateFromFile.toString()
 
-export const wasmHelperPlugin = (config: ResolvedConfig): Plugin => {
-  if (config.isBundled && config.nativePluginEnabledLevel >= 1) {
-    return nativeWasmHelperPlugin({
-      decodedBase: config.decodedBase,
-    })
-  }
-
+export const wasmHelperPlugin = (): Plugin => {
   return {
     name: 'vite:wasm-helper',
 
@@ -150,28 +139,6 @@ export default ${wasmHelperCode}
       } else {
         return null
       }
-    },
-  }
-}
-
-export const wasmFallbackPlugin = (config: ResolvedConfig): Plugin => {
-  if (config.nativePluginEnabledLevel >= 1) {
-    return nativeWasmFallbackPlugin()
-  }
-
-  return {
-    name: 'vite:wasm-fallback',
-
-    load: {
-      filter: { id: /\.wasm$/ },
-      handler(_id) {
-        throw new Error(
-          '"ESM integration proposal for Wasm" is not supported currently. ' +
-            'Use vite-plugin-wasm or other community plugins to handle this. ' +
-            'Alternatively, you can use `.wasm?init` or `.wasm?url`. ' +
-            'See https://vite.dev/guide/features.html#webassembly for more details.',
-        )
-      },
     },
   }
 }